java hindi iter_2

+++ src/main/java/org/636766fe1a6d9265ec017833/636766fe1a6d9265ec017833.java
@@ -14,24 +14,35 @@ public class FileUtils {
             throw new NullPointerException("File must not be null");
         }
 
-        file.deleteOnExit();
-
         if (file.isDirectory()) {
-            File[] files = file.listFiles();
-            if (files != null) {
-                for (File subFile : files) {
-                    forceDeleteOnExit(subFile);
-                }
-            }
+            deleteDirectoryOnExit(file);
+        } else {
+            file.deleteOnExit();
         }
     }
 
-    public static void main(String[] args) {
-        try {
-            File file = new File("path/to/your/file_or_directory");
-            forceDeleteOnExit(file);
-        } catch (IOException e) {
-            e.printStackTrace();
+    private static void deleteDirectoryOnExit(File directory) throws IOException {
+        if (!directory.exists()) {
+            return;
         }
+
+        if (!directory.isDirectory()) {
+            throw new IOException("Not a directory: " + directory);
+        }
+
+        File[] files = directory.listFiles();
+        if (files == null) {
+            throw new IOException("Failed to list contents of " + directory);
+        }
+
+        for (File file : files) {
+            if (file.isDirectory()) {
+                deleteDirectoryOnExit(file);
+            } else {
+                file.deleteOnExit();
+            }
+        }
+
+        directory.deleteOnExit();
     }
 }
